Updating

+++ Program.cs
@@ -7,12 +7,12 @@ Console.WriteLine("------------------------\n");
 
 // Ask the user to type the first number.
 Console.WriteLine("Type a number, and then press Enter");
-num1  = Convert.ToInt32(Console.ReadLine());
+num1  = Convert.ToDouble(Console.ReadLine());
 
 
 // Ask the user to type the second number.
 Console.WriteLine("Type another number, and the press Enter");
-num2 = Convert.ToInt32(Console.ReadLine());
+num2 = Convert.ToDouble(Console.ReadLine());
 
 // Ask the user to choose an option.
 Console.WriteLine("Choose an option from the following list:");
